refactor: 컨벤션에 맞게 수정
