+++ frontend/src/pages/Admin/EnterPhoneNumber/components/Dialpad/index.tsx
@@ -1,10 +1,10 @@
 import { useLocation, useNavigate } from 'react-router-dom';
-import Alert from '../../../../components/Alert';
-import useModal from '../../../../hooks/useModal';
-import useDialPad from '../hooks/useDialPad';
+import Alert from '../../../../../components/Alert';
+import useModal from '../../../../../hooks/useModal';
+import useDialPad from '../../hooks/useDialPad';
 import { BaseInput, Container, KeyContainer, Pad } from './style';
-import ROUTER_PATH from '../../../../constants/routerPath';
-import { PHONE_NUMBER_LENGTH } from '../../../../constants/magicNumber';
+import ROUTER_PATH from '../../../../../constants/routerPath';
+import { PHONE_NUMBER_LENGTH } from '../../../../../constants/magicNumber';
 
 export const DIAL_KEYS = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '←', '0', '입력'] as const;
